uadk: adjust the position of some functions
These functions are all operation functions for the device,
and they are only called in the device driver and do not
need to be presented to the user APP.
Therefore, move them to the internal public file wd_util.c
