+++ wd_util.c
@@ -9,9 +9,13 @@
 #include <pthread.h>
 #include <semaphore.h>
 #include <string.h>
+#include <sys/ioctl.h>
+#include <sys/mman.h>
 #include <ctype.h>
+
 #include "wd_sched.h"
 #include "wd_util.h"
+#include "wd_common.h"
 
 #define WD_ASYNC_DEF_POLL_NUM		1
 #define WD_ASYNC_DEF_QUEUE_DEPTH	1024
@@ -61,6 +65,127 @@ struct async_task_queue {
 	pthread_t tid;
 	int (*alg_poll_ctx)(__u32, __u32, __u32 *);
 };
+
+int wd_ctx_set_io_cmd(handle_t h_ctx, unsigned long cmd, void *arg)
+{
+	struct wd_ctx_h	*ctx = (struct wd_ctx_h *)h_ctx;
+
+	if (!ctx)
+		return -WD_EINVAL;
+
+	if (!arg)
+		return ioctl(ctx->fd, cmd);
+
+	return ioctl(ctx->fd, cmd, arg);
+}
+
+int wd_ctx_start(handle_t h_ctx)
+{
+	struct wd_ctx_h	*ctx = (struct wd_ctx_h *)h_ctx;
+	int ret;
+
+	if (!ctx)
+		return -WD_EINVAL;
+
+	ret = wd_ctx_set_io_cmd(h_ctx, UACCE_CMD_START, NULL);
+	if (ret)
+		WD_ERR("failed to start on %s (%d), ret = %d!\n",
+		       ctx->dev_path, -errno, ret);
+
+	return ret;
+}
+
+int wd_release_ctx_force(handle_t h_ctx)
+{
+	struct wd_ctx_h	*ctx = (struct wd_ctx_h *)h_ctx;
+	int ret;
+
+	if (!ctx)
+		return -WD_EINVAL;
+
+	ret = wd_ctx_set_io_cmd(h_ctx, UACCE_CMD_PUT_Q, NULL);
+	if (ret)
+		WD_ERR("failed to stop on %s (%d), ret = %d!\n",
+		       ctx->dev_path, -errno, ret);
+
+	return ret;
+}
+
+void *wd_ctx_mmap_qfr(handle_t h_ctx, enum uacce_qfrt qfrt)
+{
+	struct wd_ctx_h	*ctx = (struct wd_ctx_h *)h_ctx;
+	off_t off = qfrt * getpagesize();
+	size_t size;
+	void *addr;
+
+	if (!ctx || qfrt >= UACCE_QFRT_MAX || !ctx->qfrs_offs[qfrt]) {
+		WD_ERR("failed to check input ctx or qfrt!\n");
+		return NULL;
+	}
+
+	size = ctx->qfrs_offs[qfrt];
+
+	addr = mmap(0, size, PROT_READ | PROT_WRITE, MAP_SHARED, ctx->fd, off);
+	if (addr == MAP_FAILED) {
+		WD_ERR("failed to mmap, qfrt = %d, err = %d!\n", qfrt, -errno);
+		return NULL;
+	}
+
+	ctx->qfrs_base[qfrt] = addr;
+
+	return addr;
+}
+
+void wd_ctx_unmap_qfr(handle_t h_ctx, enum uacce_qfrt qfrt)
+{
+	struct wd_ctx_h	*ctx = (struct wd_ctx_h *)h_ctx;
+
+	if (!ctx || qfrt >= UACCE_QFRT_MAX)
+		return;
+
+	if (ctx->qfrs_offs[qfrt] != 0)
+		munmap(ctx->qfrs_base[qfrt], ctx->qfrs_offs[qfrt]);
+}
+
+unsigned long wd_ctx_get_region_size(handle_t h_ctx, enum uacce_qfrt qfrt)
+{
+	struct wd_ctx_h *ctx = (struct wd_ctx_h *)h_ctx;
+	if (!ctx || qfrt >= UACCE_QFRT_MAX)
+			return 0;
+	return ctx->qfrs_offs[qfrt];
+}
+
+void *wd_ctx_get_priv(handle_t h_ctx)
+{
+	struct wd_ctx_h	*ctx = (struct wd_ctx_h *)h_ctx;
+
+	if (!ctx)
+		return NULL;
+
+	return ctx->priv;
+}
+
+int wd_ctx_set_priv(handle_t h_ctx, void *priv)
+{
+	struct wd_ctx_h	*ctx = (struct wd_ctx_h *)h_ctx;
+
+	if (!ctx)
+		return -WD_EINVAL;
+
+	ctx->priv = priv;
+
+	return 0;
+}
+
+char *wd_ctx_get_api(handle_t h_ctx)
+{
+	struct wd_ctx_h	*ctx = (struct wd_ctx_h *)h_ctx;
+
+	if (!ctx || !ctx->dev)
+		return NULL;
+
+	return ctx->dev->api;
+}
 
 static void clone_ctx_to_internal(struct wd_ctx *ctx,
 				  struct wd_ctx_internal *ctx_in)
